requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
chardet==3.0.4
enum34==1.1.6
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
ipaddress==1.0.17
oauthlib==3.1.0
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.6
python-dotenv==0.10.3
python-twitter